requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy-utils
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
email-validator
httpx
pytest
pytest-asyncio
black
flake8
isort
python-dotenv
alembic